Gjør Radio mindre generell
